use numpy in quadora

+++ declafe/feature_gen/quadri/talib/ADOSCFeature.py
@@ -15,10 +15,10 @@ class ADOSCFeature(QuadriFeature):
 
   def quadrigen(self, col1: pd.Series, col2: pd.Series, col3: pd.Series,
                 col4: pd.Series) -> pd.Series:
-    return talib.ADOSC(col1,
-                       col2,
-                       col3,
-                       col4,
+    return talib.ADOSC(col1.astype(float),
+                       col2.astype(float),
+                       col3.astype(float),
+                       col4.astype(float),
                        fastperiod=self.fastperiod,
                        slowperiod=self.slowperiod)
