+++ declafe/feature_gen/quadri/talib/CDLADVANCEBLOCKFeature.py
@@ -20,7 +20,8 @@ class CDLADVANCEBLOCKFeature(QuadriFeature):
 
   def quadrigen(self, col1: pd.Series, col2: pd.Series, col3: pd.Series,
                 col4: pd.Series) -> pd.Series:
-    return talib.CDLADVANCEBLOCK(col1, col2, col3, col4)
+    return talib.CDLADVANCEBLOCK(col1.astype(float), col2.astype(float),
+                                 col3.astype(float), col4.astype(float))
 
   def _feature_name(self) -> str:
     return f"CDLADVANCEBLOCK_{self.col1}_{self.col2}_{self.col3}_{self.col4}"
