+++ declafe/feature_gen/quadri/talib/CDLDARKCLOUDCOVERFeature.py
@@ -20,10 +20,10 @@ class CDLDARKCLOUDCOVERFeature(QuadriFeature):
 
   def quadrigen(self, col1: pd.Series, col2: pd.Series, col3: pd.Series,
                 col4: pd.Series) -> pd.Series:
-    return talib.CDLDARKCLOUDCOVER(col1,
-                                   col2,
-                                   col3,
-                                   col4,
+    return talib.CDLDARKCLOUDCOVER(col1.astype(float),
+                                   col2.astype(float),
+                                   col3.astype(float),
+                                   col4.astype(float),
                                    penetration=self.penetration)
 
   def _feature_name(self) -> str:
