+++ declafe/feature_gen/quadri/talib/CDLDOJIFeature.py
@@ -15,7 +15,7 @@ class CDLDOJIFeature(QuadriFeature):
 
   def quadrigen(self, col1: pd.Series, col2: pd.Series, col3: pd.Series,
                 col4: pd.Series) -> pd.Series:
-    return talib.CDLDOJI(col1, col2, col3, col4)
+    return talib.CDLDOJI(col1.astype(float), col2.astype(float), col3.astype(float), col4.astype(float))
 
   def _feature_name(self) -> str:
     return f"CDLDOJI_{self.col1}_{self.col2}_{self.col3}_{self.col4}"
