+++ declafe/feature_gen/quadri/talib/CDLMATCHINGLOWFeature.py
@@ -21,10 +21,10 @@ class CDLMATCHINGLOWFeature(QuadriFeature):
   def quadrigen(self, col1: pd.Series, col2: pd.Series, col3: pd.Series,
                 col4: pd.Series) -> pd.Series:
     return talib.CDLMATCHINGLOW(
-        col1,
-        col2,
-        col3,
-        col4,
+        col1.astype(float),
+        col2.astype(float),
+        col3.astype(float),
+        col4.astype(float),
     )
 
   def _feature_name(self) -> str:
