+++ declafe/feature_gen/quadri/talib/CDLMATHOLDFeature.py
@@ -23,10 +23,10 @@ class CDLMATHOLDFeature(QuadriFeature):
   def quadrigen(self, col1: pd.Series, col2: pd.Series, col3: pd.Series,
                 col4: pd.Series) -> pd.Series:
     return talib.CDLMATHOLD(
-        col1,
-        col2,
-        col3,
-        col4,
+        col1.astype(float),
+        col2.astype(float),
+        col3.astype(float),
+        col4.astype(float),
         penetration=self.penetration,
     )
