+++ declafe/feature_gen/quadri/talib/MFIFeature.py
@@ -14,7 +14,8 @@ class MFIFeature(QuadriFeature):
 
   def quadrigen(self, col1: pd.Series, col2: pd.Series, col3: pd.Series,
                 col4: pd.Series) -> pd.Series:
-    return talib.MFI(col1, col2, col3, col4, self.period)
+    return talib.MFI(col1.astype(float), col2.astype(float), col3.astype(float),
+                     col4.astype(float), self.period)
 
   def _feature_name(self) -> str:
     return f"MFI_{self.period}_of_{self.col1}_{self.col2}_{self.col3}_{self.col4}"
